server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
httpx==0.27.0
SQLAlchemy==2.0.23
psycopg[binary]==3.1.12
Flask-JWT-Extended==4.6.0
llama-index==0.10.65
llama-index-embeddings-huggingface==0.2.2
sentence-transformers==2.7.0
pypdf==4.2.0
odfpy==1.4.1
chromadb==0.4.24
llama-index-vector-stores-chroma==0.1.4